requirements.txt
python_frontmatter==1.1.0
Requests==2.32.3
